- moved code for queue handling to consts.py for now
